% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time.R
\name{add_time}
\alias{add_time}
\title{Add time to motion detection table}
\usage{
add_time(path, exec = "exiftool")
}
\arguments{
\item{path}{a character string specifying the path of the motion detection project with folders being MotionMeerkat folders}

\item{exec}{a character string with the path efixtool (Windows, e.g. C:/Programme/exiftool/exiftool_-k_.exe)}
}
\description{
Add time to motion detection table
}
