% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_polygon.R
\name{define_polygon}
\alias{define_polygon}
\title{Retrieve coordintas for user-defined target polygon in frame}
\usage{
define_polygon(x, n)
}
\arguments{
\item{x}{an object of class 'cimg'}

\item{n}{integer number of polygon points}
}
\description{
Retrieve coordintas for user-defined target polygon in frame
}
