% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.animal.R
\name{is.animal}
\alias{is.animal}
\title{Displays MotionMeerkat hit frame and waits for user input}
\usage{
is.animal(path, frame_nr, coord, autorange = TRUE, mag = c(50, 50),
  query.text)
}
\description{
Displays MotionMeerkat hit frame and waits for user input
}
