% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.efixtool.R
\name{R.exiftool}
\alias{R.exiftool}
\title{Interface to the efixtool software to open meta data of image files}
\usage{
R.exiftool(image.path, exec = "exiftool")
}
\arguments{
\item{image.path}{a character string specifying a path to an image file}

\item{exec}{a character string with the path efixtool (Windows, e.g. C:/Programme/exiftool/exiftool_-k_.exe)}
}
\value{
data.frame with efixtool output
}
\description{
Interface to the efixtool software to open meta data of image files
}
\references{
\url{https://www.sno.phy.queensu.ca/~phil/exiftool/}
}
