% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.motion.table.R
\name{read_motion_table}
\alias{read_motion_table}
\title{Read motion detection table}
\usage{
read_motion_table(path)
}
\arguments{
\item{path}{a character string specifying the path of the motion detection project with folders being MotionMeerkat folders}
}
\value{
data.frame
}
\description{
Read motion detection table
}
