% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_animal.R
\name{remove_non_animal}
\alias{remove_non_animal}
\title{Second step after MotionMeerkat}
\usage{
remove_non_animal(path, batch = 1000,
  query.text = "Animal? (no = 0/ yes = 1/ unclear = 2); Consider top left red square!")
}
\arguments{
\item{path}{a character string specifying the path of the motion detection project with folders being MotionMeerkat folders}

\item{batch}{number of frames after which to write to file}

\item{query.text}{a character string, e.g. "Animal? (no = 0/ yes = 1/ unclear = 2)"}
}
\value{
does not return, but saves to project folder, see \link{read.motion.table}
}
\description{
Second step after MotionMeerkat
}
