% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_target.R
\name{remove_non_target}
\alias{remove_non_target}
\title{Frist step after MotionMeerkat: Delete non-target hits}
\usage{
remove_non_target(path, n)
}
\arguments{
\item{path}{a character string specifying the path of the motion detection project with folders being MotionMeerkat folders}
}
\value{
does not return, but saves to project folder, see \link{read.motion.table}
}
\description{
Removes putative movement detected by MotionMeerkat which is outside the user-devinded target area (e.g. flower)
}
