% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_coords.R
\name{target_coords}
\alias{target_coords}
\title{Extract coordinates from unser-defined target polygon}
\usage{
target_coords(coord, x)
}
\arguments{
\item{coord}{output of \link{define_polygon}}

\item{x}{an object of class 'cimg'}
}
\value{
list of points outside and inside of polygon
}
\description{
Extract coordinates from unser-defined target polygon
}
